backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.20
chromadb==0.6.3
sentence-transformers==3.4.1
transformers==4.48.1
torch==2.5.1
Pillow==11.1.0
neo4j==5.27.0
google-generativeai==0.8.4
